spliting the code, add handle for horse riding
